add index setting index_field_presence_json
By default, this setting is false, which means that an 'index update' is
required on every index for the change to take effect.

Even though '_field_presence_json' field is internal, it can be controlled by
the user through the exposed boolean flag 'index_field_presence_json'. This flag
plays a key role because it drives updates to the 'doc_mapping_uid'.

By updating 'doc_mapping_uid', we avoid merging conflicting splits, which serves
as a group-by key in the merge pipeline. As a result, only splits that share the
same doc_mapping_uid are merged together.